+++ addon/models/new-platform-flexberry-services-lock.js
@@ -1,0 +1,27 @@
+/**
+  @module ember-flexberry
+*/
+
+import DS from 'ember-data';
+import BaseModel from './base';
+import { Projection } from 'ember-flexberry-data';
+
+let Model = BaseModel.extend({
+  lockKey: DS.attr('string'),
+  userName: DS.attr('string'),
+  lockDate: DS.attr('date'),
+});
+
+Model.defineProjection('LockL', 'new-platform-flexberry-services-lock', {
+  lockKey: Projection.attr('Lock key'),
+  userName: Projection.attr('User name'),
+  lockDate: Projection.attr('Lock date'),
+});
+
+/**
+  Model lock, use in {{#crossLink "LockRouteMixin"}}{{/crossLink}}.
+
+  @class NewPlatformFlexberryServicesLockModel
+  @extends BaseModel
+*/
+export default Model;