+++ app/routes/new-platform-flexberry-services-lock-list.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry/routes/new-platform-flexberry-services-lock-list';