+++ tests/unit/models/new-platform-flexberry-services-lock-test.js
@@ -1,0 +1,8 @@
+import { moduleForModel, test } from 'ember-qunit';
+
+moduleForModel('new-platform-flexberry-services-lock', 'Unit | Model | new-platform-flexberry-services-lock');
+
+test('it exists', function(assert) {
+  let model = this.subject();
+  assert.ok(!!model);
+});